Move dashboard as cross-cutting feature module
