+++ client/src/features/dashboard/i18n/index.ts
@@ -1,0 +1,4 @@
+import es from './es.json'
+import en from './en.json'
+
+export { es as dashboardEs, en as dashboardEn }
