+++ client/src/features/dashboard/routes.tsx
@@ -1,0 +1,8 @@
+import { Route } from 'react-router-dom'
+import { Dashboard } from './pages/Dashboard'
+
+export const dashboardRoutes = (
+  <>
+    <Route path="/" element={<Dashboard />} />
+  </>
+)
